% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCorpIdent.R
\name{addLogo}
\alias{addLogo}
\title{Add INWT Logo}
\usage{
addLogo(position = getOption("ggCorpIdentLogoPosition"), alpha = 0.1,
  size = getOption("ggCorpIdentLogoSize"))
}
\arguments{
\item{position}{character; Position des Logos (s. Details)}

\item{alpha}{numeric; Transparenz (s. \code{\link[grDevices]{rgb}})}

\item{size}{numeric; Werte zwischen [0; 1]. Ein Wert von 1 entspricht der
maximal möglichen Ausdehung, so dass sich das Logo (inkl. etwas Rand) über den
maximalen Bereich der Grafik erstreckt, ohne abgeschnitten zu werden.}
}
\description{
Einbinden des INWT Logos als Wasserzeichen in ggplot Grafiken. \strong{Wichtig:}
Die Funktion \code{addLogo} sollte direkt im Anschluss an den Aufruf von
\code{\link[ggplot2]{ggplot}} erfolgen (s. Beispiele), damit das Logo als Layer hinter den Daten
eingebunden wird und nicht die Daten überlagert.
}
\details{
The position may be specified by keywords from the list \code{"default"},
\code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"},
\code{"topleft"}, \code{"top"}, \code{"topright"}, \code{"right"}, \code{"center"}
and \code{"full"}. Setting \code{"full"} corresponds to \code{addLogo(position = "full", size = 1)}.
\code{"deafult")} ist die Verwendung des Logos wie auf der Vorderseite unserer Visitenkarten.
}
