% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCorpIdent.R
\name{ggCorpIdent}
\alias{ggCorpIdent}
\title{Set Corporate Identity Theme for ggplot2 graphics}
\usage{
ggCorpIdent(base_size = 12, base_family = "Lato",
  textColor = "#2b4894", panelBackgroundColor = "#d4dbde",
  tickColor = "white", legendKeyColor = "white",
  panelGridMajorColor = "white", stripTextColor = "white",
  colors = c("#2b4894", "#068587", "#F2B134", "#ED553B", "#4FB99F"),
  logo = NULL, logoPosition = "bottomright", logoSize = 1,
  logoTransparency = 0.2)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{textColor}{text color for axis, labs, title and subtitle}

\item{panelBackgroundColor}{color for panel background}

\item{tickColor}{color for ticks}

\item{legendKeyColor}{color for legend key}

\item{panelGridMajorColor}{color for panel grid major}

\item{stripTextColor}{color for strip text}

\item{colors}{vector with corporate colors used for geoms; the colors are used
according their position in \code{colors}; if mor colors are required, values 
of \code{colors} will be interpolated via \code{\link[grDevices:colorRamp]{colorRampPalette}}.}

\item{logo}{filename for logo; the file has to be \code{PNG}.}

\item{logoPosition}{location of logo. See 'Details'.}

\item{logoSize}{numeric for logo rescaling}

\item{logoTransparency}{numeric alpha value to adjust logo transparency}
}
\description{
Sets a pre-defined theme as the standard ggplot theme via
\code{\link[ggplot2:theme_get]{theme_set}} and \code{\link[ggplot2:ggtheme]{theme_bw}}. In addtion 
to the specification of colors and fonts, a corporate logo (format: PNG) can be 
specified to serve as watermark in ggplot graphics.
}
\details{
The location may be specified by a single keyword from the 
list \code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"}, 
\code{"topleft"}, \code{"top"}, \code{"topright"}, \code{"right"} and \code{"center"}.
}
\examples{
\dontrun{
library(ggCorpIdent)
setThemeGgplot2()

# default 'ggCorpIdent' settings:
ggplot(iris,
       aes(x = Sepal.Width, y = Petal.Width, col = Species)) +
  geom_point()

# add corporate logo
logoFromPackage <- system.file("logo/logo_INWT.png", package = "ggCorpIdent", mustWork = TRUE)
ggCorpIdent(logo = logoFromPackage,
            logoSize = 0.4)
ggplot(iris,
       aes(x = Sepal.Width, y = Petal.Width, col = Species)) +
  geom_point()

# change colors and font
ggCorpIdent(base_family = "Courier",
            colors = c("dodgerblue4", "violetred1", "mintcream"),
            logo = logoFromPackage,
            logoSize = 0.4) 
ggplot(iris,
       aes(x = Sepal.Width, y = Petal.Width, col = Species)) +
  geom_point()
}
}
