% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCorpIdent.R
\name{ggplot}
\alias{ggplot}
\title{Custom ggplot function}
\usage{
ggplot(..., logo = getOption("ggCorpIdentLogoFilename"),
  logoParams = list())
}
\arguments{
\item{...}{arguments to be passed to \code{\link[ggplot2]{ggplot}}}

\item{logo}{logical: show logo?}

\item{logoParams}{list: arguments passed to \code{\link{addLogo}}}
}
\description{
Works exactly like \code{\link[ggplot2]{ggplot}} but adds a
logo via \code{\link{addLogo}} by default. If you don't want a logo, use
\code{logo = FALSE}.
}
