% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCorpIdent.R
\name{paletteFun}
\alias{paletteFun}
\title{color palette}
\usage{
paletteFun(x)
}
\arguments{
\item{x}{integer: desired number of colors}
}
\description{
Vector of color codes. If the number of requested colors
exceeds the number of available colors, interpolated colors are added.
}
