% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCorpIdent.R
\name{scale_colour_discrete}
\alias{scale_colour_discrete}
\title{Evenly spaced custom colors for discrete data}
\usage{
scale_colour_discrete(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[ggplot2]{discrete_scale}}}
}
\description{
Default color scale for categorical variables with custom colors.
}
