% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCorpIdent.R
\name{scale_fill_discrete}
\alias{scale_fill_discrete}
\title{Evenly spaced INWT colours for discrete data}
\usage{
scale_fill_discrete(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[ggplot2]{discrete_scale}}}
}
\description{
Default color scale for categorical variables with INWT colors.
}
